Modified to follow reviewer comments.
