+++ java/java.hints.declarative/src/org/netbeans/modules/java/hints/declarative/conditionapi/Matcher.java
@@ -21,12 +21,11 @@ package org.netbeans.modules.java.hints.declarative.conditionapi;
 
 import com.sun.source.tree.Tree;
 import com.sun.source.util.TreePath;
-import com.sun.source.util.Trees;
 import org.netbeans.api.java.source.support.ErrorAwareTreePathScanner;
 import java.util.Collection;
 import java.util.LinkedList;
-import javax.lang.model.element.Element;
 import org.netbeans.api.annotations.common.NonNull;
+import org.netbeans.modules.java.hints.errors.Utilities;
 import org.netbeans.spi.java.hints.MatcherUtilities;
 
 /**
@@ -91,65 +90,8 @@ public final class Matcher {
         return result[0];
     }
 
-    @SuppressWarnings("BoxedValueEquality")
     public boolean referencedIn(@NonNull Variable variable, @NonNull Variable in) {
-        final Trees trees = ctx.ctx.getInfo().getTrees();
-        final Element e = trees.getElement(ctx.getSingleVariable(variable));
-
-        if (e == null) { //TODO: check also error
-            return false;
-        }
-
-        for (TreePath tp : ctx.getVariable(in)) {
-
-            if (e.equals(trees.getElement(tp))) {
-                return true;
-            }
-
-            boolean occurs = new ErrorAwareTreePathScanner<Boolean, Void>() {
-                private boolean found = false;
-                @Override
-                public Boolean scan(Tree tree, Void p) {
-                    if (found) {
-                        return true; // fast path
-                    }
-
-                    if (tree == null) {
-                        return false;
-                    }
-
-                    TreePath currentPath = new TreePath(getCurrentPath(), tree);
-                    Element currentElement = trees.getElement(currentPath);
-
-                    if (e.equals(currentElement)) {
-                        found = true;
-                        return true;
-                    }
-
-                    return super.scan(tree, p);
-                }
-
-                @Override
-                public Boolean reduce(Boolean r1, Boolean r2) {
-                    if (r1 == null) {
-                        return r2;
-                    }
-
-                    if (r2 == null) {
-                        return r1;
-                    }
-
-                    return r1 || r2;
-                }
-
-            }.scan(tp, null) == Boolean.TRUE;
-
-            if (occurs) {
-                return true;
-            }
-        }
-
-        return false;
+        return Utilities.isReferencedIn(ctx.ctx.getInfo(), ctx.getSingleVariable(variable), ctx.getVariable(in));
     }
 
     public boolean matchesWithBind(Variable var, String pattern) {
